move to utility file
