+++ src/hooks/useAmountValidation.ts
@@ -3,6 +3,7 @@ import { useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import { QuoteResult } from 'routes/operator';
 import { RootState } from 'store';
+import { formatMinAmount } from 'utils/formatNumber';
 import { isMinAmountError } from 'utils/sdkv2';
 
 export type AmountValidationResult = {
@@ -93,19 +94,4 @@ export const useAmountValidation = (props: Props): AmountValidationResult => {
   }
 
   return {};
-};
-
-// Minimum amounts are approximations anyway so we don't need to display ultra-precise figures here.
-// For amounts >  999, we simply round up.
-// For amounts <= 999, we use toPrecision(3) which only shows the first 3 non-zero digits.
-// This way we're not showing excessive precision for any value
-const formatMinAmount = (minAmount: sdkAmount.Amount): string => {
-  const formatted = sdkAmount.display(minAmount);
-  // Minimum amounts are approximations so we do a little floating point fudging
-  const asNumber = parseFloat(formatted);
-  if (asNumber > 999) {
-    return Math.ceil(asNumber).toString();
-  } else {
-    return asNumber.toPrecision(3);
-  }
 };
